backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
python-dotenv
